failed attempt at fastforwarding
try modifying the current frame and the frames it links to in the hope
that a return causes the interpreter to leave off where it the frames
tell it. this turns out to not work. the interpreter's state is only
partially captured in the frame objects. some of it is stored in the CPU
stack. This also explains why generators run only in their own frame
level and can't return and generate from their parent frame.