save callgraph with eval_frame
Instead of using sys.set_trace(), use
[PEP524](https://www.python.org/dev/peps/pep-0523/) eval_frame to
trace calls to functions. This is much faster than set_trace() because
it gets called once per function entry instead of for every line in the
code.

The handler is also writting in cython, so it can be optimized beyond
what the set_trace() version before could.