failed attempt at retrieving stack depth from generators
i noticed the yield instruction in the python3.9 code causes the
stacktop to get stored in the frame value. i hoped to introduce a yield
before each function call to cause the stacktop to get saved. but of
course this doesn't work because the generator gets constructed by
cython, not by frameevaldefault.