Checkpointer works on functions
New idea: instead of checkpointing iterators, checkpoint functions, as
I wanted to do initially.

The insight is that we can guess at the size of the stack by counting
the number of locals, the arguments to CALL_FUNCTION, and the number of
arguments the function consumes.

This might not always work because the python compiler can, in
principle, generate code that pushes other things on the stack before a
CALL_FUNCTION opcode beyond the ones I'm enumerating. But I haven't yet
run into such a case.

Left to do: handle argument counts in functions