Jump through exception handlers
Previous to this, calling jump() in an exception handler would cause a
SEGFAULT. This was happening because during exception handling, the
stack contains the exception triplet, which the stack saving logic
wasn't taking into account.

With this commit, look into the block stack and 3 to our estimate of the
stack depth for every exception block stack.

Now you can call jump() inside an exception handler.