Support while loops too
Previously, I relied on the PyTryBlock stack to determine the level of
nesting of for loops. This has two problems 1) while loops also push
onto this stack, but while loops do not push an iterator on the python
stack. I needed to find a way to distinguish between these loops. 2)
@user2357112supportsMonica on Stackoverflow pointed out that Python 3.8+
no longer pushes onto the PyTryBlock for loops, so this solution wasn't
going to work.

So instead, we count the level of nesting of for loops that surrounds a
call statement by looking at the bytecode.